requirements.txt
pandas
numpy
matplotlib
seaborn
faker
streamlit
fastapi
uvicorn
openai
pillow
plotly
sqlalchemy
python-dotenv
